Refactor: 코드 분리 및 클래스명 수정
- 클래스는 좀 더 세분화하여 코드 분리
- 클래스 이름 구체적으로 변경
- const TIME = 120 대신 static TIME = 120 으로 변경
